requirements.txt
requests
pytest
pytest-watch
ruff
numpy
types-requests
mypy
python-dotenv